requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
pandas
numpy
scikit-learn
plotly
pyyaml
tqdm 